pyproject.toml
[project]
name = "goose"
description = "a programming agent that runs on your machine"
version = "0.8.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "attrs>=23.2.0",
    "rich>=13.7.1",
    "ruamel-yaml>=0.18.6",
    "exchange>=0.7.6",
    "click>=8.1.7",
    "prompt-toolkit>=3.0.47",
]
author = [{ name = "Block", email = "ai-oss-tools@block.xyz" }]
packages = [{ include = "goose", from = "src" }]

[project.entry-points."goose.toolkit"]
developer = "goose.toolkit.developer:Developer"
github = "goose.toolkit.github:Github"
screen = "goose.toolkit.screen:Screen"
repo_context = "goose.toolkit.repo_context.repo_context:RepoContext"

[project.entry-points."goose.profile"]
default = "goose.profile:default_profile"

[project.entry-points."goose.command"]
file = "goose.command.file:FileCommand"

[project.entry-points."goose.cli"]
goose = "goose.cli.main:goose_cli"

[project.scripts]
goose = "goose.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.2",
    "codecov>=2.1.13",
]

